+++ src/redux/sagas/expenditures/editExpenditure.ts
@@ -1,6 +1,5 @@
 import { call, fork, put, takeEvery } from 'redux-saga/effects';
 import { ClientType } from '@colony/colony-js';
-import { BigNumber } from 'ethers';
 
 import { Action, ActionTypes, AllActions } from '~redux';
 import { ExpenditurePayoutFieldValue } from '~common/Expenditures/ExpenditureForm';
@@ -9,7 +8,7 @@ import {
   putError,
   getSetExpenditureValuesFunctionParams,
   initiateTransaction,
-  getPayoutsWithSlotIds,
+  getResolvedExpenditurePayouts,
 } from '../utils';
 import {
   createTransaction,
@@ -23,52 +22,8 @@ function* editExpenditure({
 }: Action<ActionTypes.EXPENDITURE_EDIT>) {
   const txChannel = yield call(getTxChannel, meta.id);
 
-  const payoutsWithSlotIds = getPayoutsWithSlotIds(payouts);
-  /**
-   * @NOTE: Resolving payouts means making sure that for every slot, there's only one payout with non-zero amount.
-   * This is to meet the UI requirement that there should be one payout per row.
-   */
-  const resolvedPayouts: ExpenditurePayoutFieldValue[] = [];
-
-  payoutsWithSlotIds.forEach((payout) => {
-    // Add payout as specified in the form
-    resolvedPayouts.push(payout);
-
-    const existingSlot = expenditure.slots.find(
-      (slot) => slot.id === payout.slotId,
-    );
-
-    // Set the amounts for any existing payouts in different tokens to 0
-    resolvedPayouts.push(
-      ...(existingSlot?.payouts
-        ?.filter(
-          (slotPayout) =>
-            slotPayout.tokenAddress !== payout.tokenAddress &&
-            BigNumber.from(slotPayout.amount).gt(0),
-        )
-        .map((slotPayout) => ({
-          slotId: payout.slotId,
-          recipientAddress: payout.recipientAddress,
-          tokenAddress: slotPayout.tokenAddress,
-          amount: '0',
-          claimDelay: payout.claimDelay,
-        })) ?? []),
-    );
-  });
-
-  // If there are now less payouts than expenditure slots, we need to remove them by setting their amounts to 0
-  const remainingSlots = expenditure.slots.slice(payouts.length);
-  remainingSlots.forEach((slot) => {
-    slot.payouts?.forEach((payout) => {
-      resolvedPayouts.push({
-        slotId: slot.id,
-        recipientAddress: slot.recipientAddress ?? '',
-        tokenAddress: payout.tokenAddress,
-        amount: '0',
-        claimDelay: slot.claimDelay ?? 0,
-      });
-    });
-  });
+  const resolvedPayouts: ExpenditurePayoutFieldValue[] =
+    getResolvedExpenditurePayouts(expenditure, payouts);
 
   try {
     yield fork(createTransaction, meta.id, {
